refactor(version): move semver precedence compare into internal/version
Review #4969353348 (r3810812621): the comparators outgrew the proxy —
internal/version parsed semver but could not order it, so Compare fills
a package gap instead of living as cmd/gortex-private helpers. The
§11 precedence test moves with it (TestComparePrecedence); the skew
remedy stays in cmd/gortex as CLI messaging.
