+++ cmd/gortex/proxy.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/zzet/gortex/internal/daemon"
@@ -176,81 +174,13 @@ func daemonSkewWarning(daemonVer, localVer string) string {
 	if dErr != nil || lErr != nil {
 		return base + " — run 'gortex daemon restart' or 'gortex upgrade'"
 	}
-	switch compareSemver(d, l) {
+	switch semver.Compare(d, l) {
 	case -1: // daemon older — restarting respawns it from this newer binary
 		return base + " — run 'gortex daemon restart' to upgrade the daemon"
 	case 1: // daemon newer — this binary is the stale side
 		return base + " — this binary is older than the running daemon — run 'gortex upgrade'"
 	default: // same precedence, different build metadata
 		return base + " — run 'gortex daemon restart' or 'gortex upgrade'"
-	}
-}
-
-// compareSemver orders two Versions by SemVer 2.0.0 precedence:
-// major.minor.patch numerically, then pre-release (absent outranks
-// present; identifiers compared per comparePrereleaseIdent). Build
-// metadata is ignored — per the spec it never affects precedence.
-// Returns -1, 0, or +1.
-func compareSemver(a, b semver.Version) int {
-	switch {
-	case a.Major != b.Major:
-		return compareInt(a.Major, b.Major)
-	case a.Minor != b.Minor:
-		return compareInt(a.Minor, b.Minor)
-	case a.Patch != b.Patch:
-		return compareInt(a.Patch, b.Patch)
-	case a.Prerelease == b.Prerelease:
-		return 0
-	case a.Prerelease == "":
-		return 1 // the release outranks any of its pre-releases
-	case b.Prerelease == "":
-		return -1
-	default:
-		return comparePrerelease(a.Prerelease, b.Prerelease)
-	}
-}
-
-// comparePrerelease compares two dot-separated pre-release identifier
-// lists per SemVer 2.0.0: identifier by identifier, and once every
-// shared identifier is equal, the shorter list ranks below the longer.
-func comparePrerelease(a, b string) int {
-	as := strings.Split(a, ".")
-	bs := strings.Split(b, ".")
-	for i := 0; i < len(as) && i < len(bs); i++ {
-		if c := comparePrereleaseIdent(as[i], bs[i]); c != 0 {
-			return c
-		}
-	}
-	return compareInt(len(as), len(bs))
-}
-
-// comparePrereleaseIdent compares one pre-release identifier pair:
-// numeric identifiers compare numerically and rank below alphanumeric
-// ones; alphanumeric identifiers compare in ASCII sort order.
-func comparePrereleaseIdent(a, b string) int {
-	an, aErr := strconv.Atoi(a)
-	bn, bErr := strconv.Atoi(b)
-	switch {
-	case aErr == nil && bErr == nil:
-		return compareInt(an, bn)
-	case aErr == nil:
-		return -1 // numeric identifiers rank below alphanumeric
-	case bErr == nil:
-		return 1
-	default:
-		return strings.Compare(a, b)
-	}
-}
-
-// compareInt is the three-way integer compare the semver helpers use.
-func compareInt(a, b int) int {
-	switch {
-	case a < b:
-		return -1
-	case a > b:
-		return 1
-	default:
-		return 0
 	}
 }
